+++ Core/Net/NetFileCache.cs
@@ -135,7 +135,9 @@ namespace CKAN
         /// <summary>>
         /// Returns the filename of an already cached url or null otherwise
         /// </summary>
-        public string GetCachedFilename(Uri url)
+        /// <param name="url">The URL to check for in the cache</param>
+        /// <param name="remoteTimestamp">Timestamp of the remote file, if known; cached files older than this will be considered invalid</param>
+        public string GetCachedFilename(Uri url, DateTime? remoteTimestamp = null)
         {
             log.DebugFormat("Checking cache for {0}", url);
 
@@ -169,7 +171,19 @@ namespace CKAN
                 string filename = Path.GetFileName(file);
                 if (filename.StartsWith(hash))
                 {
-                    return file;
+                    // Check local vs remote timestamps; if local is older, then it's invalid.
+                    // null means we don't know the remote timestamp (so file is OK)
+                    if (remoteTimestamp == null
+                        || remoteTimestamp < File.GetLastWriteTime(file).ToUniversalTime())
+                    {
+                        // File not too old, use it
+                        return file;
+                    }
+                    else
+                    {
+                        // Local file too old, delete it
+                        File.Delete(file);
+                    }
                 }
             }
 
@@ -209,9 +223,9 @@ namespace CKAN
         }
 
         /// <summary>
-        /// Check whether a ZIP file is validation
+        /// Check whether a ZIP file is valid
         /// </summary>
-        /// <param name="filename">path to zip file to check</param>
+        /// <param name="filename">Path to zip file to check</param>
         /// <param name="invalidReason">Description of problem with the file</param>
         /// <returns>
         /// True if valid, false otherwise. See invalidReason param for explanation.
@@ -224,15 +238,27 @@ namespace CKAN
                 {
                     using (ZipFile zip = new ZipFile(filename))
                     {
-                        // Perform CRC check.
-                        if (zip.TestArchive(true))
+                        string zipErr = null;
+                        // Perform CRC and other checks
+                        if (zip.TestArchive(true, TestStrategy.FindFirstError,
+                            (TestStatus st, string msg) =>
+                            {
+                                // This delegate is called as TestArchive proceeds through its
+                                // steps, both routine and abnormal.
+                                // The second parameter is non-null if an error occurred.
+                                if (st != null && !st.EntryValid && !string.IsNullOrEmpty(msg))
+                                {
+                                    // Capture the error string so we can return it
+                                    zipErr = $"Error in step {st.Operation} for {st.Entry?.Name}: {msg}";
+                                }
+                            }))
                         {
                             invalidReason = "";
                             return true;
                         }
                         else
                         {
-                            invalidReason = "ZipFile.TestArchive(true) returned false";
+                            invalidReason = zipErr ?? "ZipFile.TestArchive(true) returned false";
                             return false;
                         }
                     }